feat: include minerId in task definition
When checking whether a measurement is for a valid round task, compare
both `cid` and `minerId` fields

When picking a checker node in an IPv4 subnet to receive the reward,
include `minerId` in the grouping key: When two nodes in the same subnet
retrieve the same CID but from different miners, they should be both
rewarded.

In practice, this second change will manifest very infrequently, because
it's unprobable to have a single round with two tasks for the same CID.
As a result, I did not need to make any changes to the integration
test running against real measurements.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>