fix: calculate committee stats from honest measurements
When calculating number of participants/subnets/measurements per
committee (per retrieval task), we must use only accepted measurements.

When using all measurements, we include data from measurements for tasks
(committees) that are not valid for this round, which distorts the
statistics. For example, the minimum number of participants per
committee is much smaller.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>