feat: form committees and find majority results
Group accepted measurements to committees on a per-retrieval-task basis.
Evaluate each committee to find an absolute majority result. Reject
results that are in minority.

When the committee is too small to give us confidence in a majority
being honest, or if we cannot find an absolute majority, then reject
all measurements in such committee.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>